+++ core/fiber/shared/resume.rb
